new_variant
x

+++ const.h
@@ -1,9 +1,9 @@
 #ifndef CONST_H_INCLUDED
 #define CONST_H_INCLUDED
 
-int type1=0x12;//ON
-int type2=0x13;//OFF
-int type3=0x20;//COLOR
+double type1=0x12;//ON
+double type2=0x13;//OFF
+double type3=0x20;//COLOR
 
 
 #endif // CONST_H_INCLUDED
